Refactor: moved member look up functions to shared_util module
getMemberById and getMemberByEmail are now accessible thru a shared 'members' module. This prevents duplicate code for lambdas that need to look up members (for eg. getMembers and assignPayment)
