+++ infra/lambdas/assigned_payments/assignPayment/index.js
@@ -1,10 +1,6 @@
-const { DynamoDBClient } = require("@aws-sdk/client-dynamodb");
-const {
-    QueryCommand,
-    DynamoDBDocumentClient
-} = require("@aws-sdk/lib-dynamodb");
-
 const { getSupabase } = require("../../shared_utils/supabase");
+const { getMemberById } = require("../../shared_utils/members");
+
 
 const PAYMENTS_TABLE = "Payments";
 const PAYMENT_ID_ATTR = "payment_id";
@@ -12,11 +8,8 @@ const ASSIGNED_PAYMENTS_TABLE = "AssignedPayments";
 
 const SUPABASE_SECRET_ID = process.env.SUPABASE_SECRET_ID;
 const REGION = process.env.AWS_REGION;
-const ddb = DynamoDBDocumentClient.from(new DynamoDBClient({ region : REGION }));
 
 const MEMBER_ID_ATTR = "member_id";
-const MEMBERS_TABLE  = "members";
-
 const REQUIRED_FIELDS = ["member_id", "payment_id", "assigned_on", "status"];
 
 function dummyCognito(){
@@ -50,7 +43,7 @@ exports.handler = async (event) => {
         const memberId = parseInt(payload[MEMBER_ID_ATTR]);
         const paymentId = parseInt(payload[PAYMENT_ID_ATTR]);
 
-        const member = await getMemberByMemberId(memberId);
+        const member = await getMemberById(memberId);
         if(member.length == 0){
             return { statusCode: 400, body: "Invalid member ID." };
         }
@@ -91,27 +84,4 @@ exports.handler = async (event) => {
             body : JSON.stringify({ error : err.message })
         };
     }
-}
-
-/**
- * TODO: this is copy pasted from getMembers. We should probably create a shared library for this.
- * @param {*} memberId 
- * @returns 
- */
-async function getMemberByMemberId(memberId) {
-  const responseItems = [];
-  let lastKey;
-  do {
-    const response = await ddb.send(new QueryCommand({
-      TableName: MEMBERS_TABLE,
-      KeyConditionExpression: "#e = :e",
-      ExpressionAttributeNames:  { "#e": MEMBER_ID_ATTR },
-      ExpressionAttributeValues: { ":e": memberId },
-      ExclusiveStartKey: lastKey,
-    }));
-    responseItems.push(...(response.Items || []));
-    lastKey = response.LastEvaluatedKey;
-  } while (lastKey);
-
-  return responseItems;
 }
